Se refactorizaron varias cosas en frontend, para mayor entendimiento Se volvió a mejorar flujo de autenticación, de tal forma que ahora cuando se expire el token, redirije a la home. Además, se encuentra en el proceso de integrar roles. Se agregaron endpoints en backend para eliminar sedes y eventos, así como también se cambió la UI de inventario para acoplar estos endpoints TO DO: RequireRole, y empezar con la página de eventos, vista de usuario normal

+++ bookingservice/src/main/java/com/nicog/bookingservice/event/BookingEvent.java
@@ -1,6 +1,5 @@
 package com.nicog.bookingservice.event;
 
-import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -15,5 +14,5 @@ public class BookingEvent {
     private String userId;
     private Long eventId;
     private Long ticketCount;
-    private BigDecimal totalPrice;
+    private Double totalPrice;
 }
