+++ inventoryservice/src/main/java/com/nicog/inventoryservice/request/CreateEventRequest.java
@@ -1,6 +1,5 @@
 package com.nicog.inventoryservice.request;
 
-import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -15,5 +14,5 @@ public class CreateEventRequest {
     private String name;
     private Long totalCapacity;
     private Long venueId;
-    private BigDecimal ticketPrice;
+    private Double ticketPrice;
 }
